requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
lxml>=4.9.0
textblob>=0.17.0
# Optional packages - install manually if needed
# transformers>=4.30.0
# torch>=2.0.0
# openai>=1.0.0
# requests-html>=0.10.0
# selenium>=4.15.0
# webdriver-manager>=4.0.0
# spacy>=3.6.0
